Activity_Zones and all system of zones is finished and tested, including in Cart_Shopping

+++ app/src/main/java/com/example/yaencasa/Data/ModelZone.kt
@@ -1,7 +1,7 @@
 package com.example.yaencasa.Data
 
 class ModelZone(
-    val idZone: Int,
+    val idZone: Long,
     val name: String,
-    val price: Int
+    val price: Double
 )
